librarywebsite-project/requirements.txt
Django==2.1.5
gunicorn==19.9.0
pyrebase
python-firebase==1.2
firebase-admin==2.16.0
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
